requirements.txt
setuptools==70.3.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.10.0.84
pygetwindow==0.0.9
PyAutoGUI==0.9.54
pyobjc==10.3.1
pyobjc-core==10.3.1
paddleocr==2.8.0
paddlepaddle==3.0.0b1
Levenshtein==0.25.1
pytest==8.2.2
selenium==4.23.1
psutil==6.0.0
